requirements.txt
PyQt6>=6.5.0
pycryptodome>=3.19.0
bcrypt>=4.1.0
flask>=3.0.0
requests>=2.31.0
